Move _accessors and values.
